+++ owsi-core-components/owsi-core-component-jpa-more/src/main/java/fr/openwide/core/jpa/more/business/task/service/IQueuedTaskHolderService.java
@@ -3,7 +3,7 @@ package fr.openwide.core.jpa.more.business.task.service;
 import fr.openwide.core.jpa.business.generic.service.IGenericEntityService;
 import fr.openwide.core.jpa.more.business.task.model.QueuedTaskHolder;
 
-public interface IQueuedTaskHolderService extends IGenericEntityService<Integer, QueuedTaskHolder> {
+public interface IQueuedTaskHolderService extends IGenericEntityService<Long, QueuedTaskHolder> {
 
 	QueuedTaskHolder getNextTaskForExecution(String taskType);
