+++ owsi-core-components/owsi-core-component-jpa-more/src/test/java/fr/openwide/core/test/jpa/more/business/entity/dao/ITestEntityDao.java
@@ -3,6 +3,6 @@ package fr.openwide.core.test.jpa.more.business.entity.dao;
 import fr.openwide.core.jpa.business.generic.dao.IGenericEntityDao;
 import fr.openwide.core.test.jpa.more.business.entity.model.TestEntity;
 
-public interface ITestEntityDao extends IGenericEntityDao<Integer, TestEntity> {
+public interface ITestEntityDao extends IGenericEntityDao<Long, TestEntity> {
 
 }
