+++ owsi-core-components/owsi-core-component-jpa-security/src/main/java/fr/openwide/core/jpa/security/business/authority/dao/IAuthorityDao.java
@@ -3,5 +3,5 @@ package fr.openwide.core.jpa.security.business.authority.dao;
 import fr.openwide.core.jpa.business.generic.dao.IGenericEntityDao;
 import fr.openwide.core.jpa.security.business.authority.model.Authority;
 
-public interface IAuthorityDao extends IGenericEntityDao<Integer, Authority> {
+public interface IAuthorityDao extends IGenericEntityDao<Long, Authority> {
 }
