+++ owsi-core-components/owsi-core-component-jpa-security/src/main/java/fr/openwide/core/jpa/security/business/person/dao/AbstractPersonGroupDaoImpl.java
@@ -5,7 +5,7 @@ import fr.openwide.core.jpa.security.business.person.model.AbstractPerson;
 import fr.openwide.core.jpa.security.business.person.model.AbstractPersonGroup;
 
 public abstract class AbstractPersonGroupDaoImpl<G extends AbstractPersonGroup<G, P>, P extends AbstractPerson<P>>
-		extends GenericEntityDaoImpl<Integer, G> implements IPersonGroupDao<G, P> {
+		extends GenericEntityDaoImpl<Long, G> implements IPersonGroupDao<G, P> {
 
 	public AbstractPersonGroupDaoImpl() {
 		super();
