+++ owsi-core-examples/owsi-core-example-jpa/src/main/java/fr/openwide/jpa/example/business/company/dao/CompanyDaoImpl.java
@@ -6,7 +6,7 @@ import fr.openwide.core.jpa.business.generic.dao.GenericEntityDaoImpl;
 import fr.openwide.jpa.example.business.company.model.Company;
 
 @Repository("companyDao")
-public class CompanyDaoImpl extends GenericEntityDaoImpl<Integer, Company> implements CompanyDao {
+public class CompanyDaoImpl extends GenericEntityDaoImpl<Long, Company> implements CompanyDao {
 
 	public CompanyDaoImpl() {
 		super();
