+++ owsi-core-examples/owsi-core-example-jpa/src/main/java/fr/openwide/jpa/example/business/project/dao/ProjectDaoImpl.java
@@ -6,7 +6,7 @@ import fr.openwide.core.jpa.business.generic.dao.GenericEntityDaoImpl;
 import fr.openwide.jpa.example.business.project.model.Project;
 
 @Repository("projectDao")
-public class ProjectDaoImpl extends GenericEntityDaoImpl<Integer, Project> implements ProjectDao {
+public class ProjectDaoImpl extends GenericEntityDaoImpl<Long, Project> implements ProjectDao {
 
 	public ProjectDaoImpl() {
 		super();
